Remove internal grpc from engine
No longer used.
