--- api/v1/inference_engine.swagger.json
@@ -1,68 +1,0 @@
-{
-  "swagger": "2.0",
-  "info": {
-    "title": "api/v1/inference_engine.proto",
-    "version": "version not set"
-  },
-  "tags": [
-    {
-      "name": "InferenceEngineInternalService"
-    }
-  ],
-  "consumes": [
-    "application/json"
-  ],
-  "produces": [
-    "application/json"
-  ],
-  "paths": {},
-  "definitions": {
-    "protobufAny": {
-      "type": "object",
-      "properties": {
-        "@type": {
-          "type": "string"
-        }
-      },
-      "additionalProperties": {}
-    },
-    "rpcStatus": {
-      "type": "object",
-      "properties": {
-        "code": {
-          "type": "integer",
-          "format": "int32"
-        },
-        "message": {
-          "type": "string"
-        },
-        "details": {
-          "type": "array",
-          "items": {
-            "$ref": "#/definitions/protobufAny"
-          }
-        }
-      }
-    },
-    "v1ListModelsResponse": {
-      "type": "object",
-      "properties": {
-        "models": {
-          "type": "array",
-          "items": {
-            "$ref": "#/definitions/v1Model"
-          }
-        }
-      }
-    },
-    "v1Model": {
-      "type": "object",
-      "properties": {
-        "id": {
-          "type": "string",
-          "description": "id is the model id."
-        }
-      }
-    }
-  }
-}
