+++ configs/engine/config.yaml
@@ -1,4 +1,3 @@
-internalGrpcPort: 8082
 ollamaPort: 8080
 objectStore:
 inferenceManagerServerWorkerServiceAddr: server:8082
